refactor(file-writer): move
