+++ src/fs/file/writer/ICodefendFileWriter.ts
